poc-python-firestore/src/functions/bigquery/requirements.txt
google-cloud-bigquery==1.20.0